Prevent the modal overlay from briefly appearing after chain is synced
After the chain is fully synced and we have the wallet open, whenever
a new block header arrives the modaloverlay will briefly appear until
the next block has been processed.  This is annoying and so we disable
the reappearance of the overlay when the chain is already or nearly
synced.