Do not assign PreferencesRole to the Unlimited dialog widget
That way we fix a layout problem on osx. Since only one widget
is going to be mapped to the "Preferences..." item on OSX menu,
in case more than one widget is assigned to that role only the last
to be instantiated on the GUI will be assigned to the "Preferences..."
and visible, all the rest won't be displayed at all.