Removed the "scheduled" stage
This is necessary because to perform certain "scheduled" tasks we need
tasks that we exeecute during normal pipelines (the one that is going
to be run when we push or made chengs to a MR).

In fact "stage" field is used not to group "similar" tasks, but to mark
tasks that could be run in parallel.