Add a new task: test-ubuntu-qa-extended
This task is going to execute only extended functional tests.
Those tests are executed when `--extended` or `--extended-only`
flags are passed to `rpc-tests.py`.

This new CI task depends on:

- build-ubuntu-deps
- build-ubuntu

The above  are run during normal "pushed" pipelines and this is the
reason why we added to the `rules` section another conditio under
which the tasks are executed, ie $EXTENDED_TESTS == "True"