Remove leftover from the SigOps counting machinery
The codebase current mitigation against DoS attack
leveraging very long and difficult to evaluate transactions
is called SigCheks. Before that another kind of mechanism
was in place, namely SigOps. If you are interested in reading
about why the change has been made have a look at:
https://upgradespecs.bitcoincashnode.org/2020-05-15-sigchecks/

This commit remove a bunch of unused constants and a function
that were part of the SigOps counting and limiting system.